Major refactor to combine Command instead of separated Type and ID.
Prior assumption of 0x33 being get is only generally true but not
necessary. Also there are way more different commands than I
originally throught. And different firmware actually has different
command sets (majority is the same).

Because of above, we had to branch logics and use firmwaree as the
first check before we use a command. This is way more unscalable
and yet necessary. It would possibly require a long time to reverse
engineer or validate at risk of setting unexpected values in EEPROM.